LayoutTests/http/tests/media/video-donottrack.html

<!DOCTYPE html>
<html>
    <head>
        <script src=../../media-resources/video-test.js></script>
        <script src=../../media-resources/media-file.js></script>
        <script>
            function runTest () {
                if (window.testRunner)
                    window.testRunner.setShouldSendDoNotTrackHTTPHeader(true);

                findMediaElement();
                var movie = findMediaFile('video', 'test');
                var type = mimeTypeForExtension(movie.split('.').pop());

                source = document.getElementById('source');
                source.src = 'http://127.0.0.1:8000/media/resources/video-donottrack-check-donottrack.php?name=' + movie + '&type=' + type;
                source.type = type;

                waitForEventAndFail('error');
                waitForEventAndEnd('canplay');
                video.load();
            }
        </script>
    </head>
    <body onload="runTest()">
        <video id="video">
            <source id="source">
        </video>
        Tests that the media player will send the relevant referer when requesting the media file.<br/>
    </body>
</html>
