LayoutTests/http/tests/media/video-donottrack-expected.txt

 Tests that the media player will send the relevant referer when requesting the media file.
EVENT(canplay)
END OF TEST
